requirements.txt
appdirs==1.4.3
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
DictObject==0.1.3
future==0.16.0
idna==2.6
luckydonald-utils==0.57
oauthlib==2.0.6
pyaes==1.6.1
pyasn1==0.4.2
PySocks==1.6.8
pyTelegramBotAPI==3.6.2
pytg==0.4.10
python-telegram-bot==10.0.1
pytz==2018.3
PyYAML==3.11
requests>=2.11.1
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
python-telegram-bot==10.0.1
telegram-send==0.19
Telethon==0.18.2.2
tweepy==3.6.0
typing==3.6.4
tzlocal==1.5.1
urllib3==1.22
virtualenv==15.0.1
